+++ libs/openai/src/models/engine.model.ts
@@ -1,23 +1,25 @@
 export enum EngineName {
   Davinci = 'davinci',
+  DavinciInstructBeta = 'davinci-instruct-beta ',
   Curie = 'curie',
+  CurieInstructBeta = 'curie-instruct-beta',
   Babbage = 'babbage',
-  Ada = 'ada'
+  Ada = 'ada',
 }
 
 export interface Engine {
-  'id': EngineName
-  'object': 'engine'
-  'created': null | string
-  'max_replicas': null | string
-  'owner': 'openai',
-  'permissions': null | string
-  'ready': boolean,
-  'ready_replicas': null | string
-  'replicas': null | string
+  id: EngineName
+  object: 'engine'
+  created: null | string
+  max_replicas: null | string
+  owner: 'openai'
+  permissions: null | string
+  ready: boolean
+  ready_replicas: null | string
+  replicas: null | string
 }
 
 export interface ListEngine {
-  'data': Engine[],
-  'object': 'list'
+  data: Engine[]
+  object: 'list'
 }